LayoutTests/media/media-source/media-source-real-canplaythrough.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-canplaythrough</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Real-MP4 equivalent of media-source-canplaythrough.html
    // Tests that loadedmetadata → loadeddata → canplay → canplaythrough events
    // fire after appending enough data.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();

        waitForEvent('loadedmetadata');
        waitForEvent('loadeddata');
        waitForEvent('canplay');
        waitForEventAndEnd('canplaythrough');

        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // Append a large block of data to trigger canplaythrough.
        // Using all sync frames so the decoder can process them.
        var samples = [];
        for (var i = 0; i < 100; i++) {
            samples.push({
                duration: 1000,
                compositionTimeOffset: 0,
                isSync: true
            });
        }
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: samples }]
        });
        sourceBuffer.appendBuffer(media);
        // Test ends via canplaythrough event handler (waitForEventAndEnd)
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
