LayoutTests/media/media-source/media-source-real-canplaythrough-expected.txt


EVENT(sourceopen)
EVENT(loadedmetadata)
EVENT(updateend)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
END OF TEST
